scheduler/requirements.txt
backoff==1.10.0
requests==2.25.0
python-decouple==3.3
